+++ apps/express/src/index.ts
@@ -10,7 +10,7 @@ const port = 3000;
 const cache = new MemoryCache();
 const sequelize = new Sequelize({
   dialect: "sqlite",
-  storage: "./lens.db",
+  storage: "./random.db",
   benchmark: true,
   logQueryParameters: true,
   logging: (sql, timing) => {
@@ -133,7 +133,7 @@ app.get("/throw-error", async (_, res) => {
   new MyRandomClass().throwsErrors();
 });
 
-handleExceptions(app);
+handleExceptions();
 
 app.listen(port, () => {
   console.log(`Server is running at http://localhost:${port}`);