+++ packages/nestjs/lib/adapters/fastify.ts
@@ -1,0 +1,3 @@
+import { FastifyAdapter } from "@lensjs/fastify";
+
+export default class FastifyNestAdapter extends FastifyAdapter {}