+++ publish.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -e
+
+npx changeset init
+npx changeset 
+npx changeset publish