+++ packages/express/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @lensjs/express
 
+## 1.3.2
+
+### Patch Changes
+
+- Fix Express Static Server
+
+## 1.3.1
+
+### Patch Changes
+
+- Drop Using Emittery
+- Updated dependencies
+  - @lensjs/core@2.2.1
+  - @lensjs/watchers@1.0.17
+
+## 1.3.0
+
+### Minor Changes
+
+- Add Fastify And NestJs Adapters
+
+### Patch Changes
+
+- Updated dependencies
+  - @lensjs/core@2.2.0
+  - @lensjs/watchers@1.0.16
+
 ## 1.2.1
 
 ### Patch Changes