+++ packages/handlers/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lensjs/watchers
 
+## 1.0.17
+
+### Patch Changes
+
+- Drop Using Emittery
+- Updated dependencies
+  - @lensjs/core@2.2.1
+
+## 1.0.16
+
+### Patch Changes
+
+- Add Fastify And NestJs Adapters
+- Updated dependencies
+  - @lensjs/core@2.2.0
+
 ## 1.0.15
 
 ### Patch Changes